Create Code
RSA is very high speed of encryption .
Your sensitive details like your bank details are now secured using RSA.
Now the bank transactions can be done securely without worrying about attacker getting access to the database as the data will be in encrypted form.
RSA involves a public key and a private key. The public key can be known by everyone and is used for encrypting messages. Messages encrypted with the public key can only  be decrypted in a reasonable amount of time using the private key. The keys for the  RSA algorithm are generated the following way:
Choose two distinct prime number p and q.
For security purposes, the integers p and q should be chosen at random, and should be of similar bit-length. Prime integers can be efficiently found using a primality test.
Compute n = pq.
n is used as the modulus for both the public and private keys. Its length, usually expressed in bits, is the key length.
Computeφ(n)= φ(p)φ(q)=(p−1)(q−1)=n -(p+ q-1),whereφisEuler’stotient function.
Choose an integer e such that 1 < e <φ(n) and gcd(e, φ(n)) = 1; i.e., e and φ(n) are coprime. e is released as the public key exponent.
e having a short bit-length and small Hamming weight results in more efficient encryption – most commonly 216 + 1 = 65,537. However, much smaller values of e (such as 3) have been shown to be less secure in some settings.
Determine d as d ≡e−1 (mod φ(n)); i.e., d is the multiplicative inverse of e (modulo
φ(n)).
This is more clearly stated as: solve for d given d⋅e ≡1 (mod φ(n))
This is often computed using the extended Euclidean algorithm. Using the pseudocode in the Modular integers section, inputs a and n correspond to e and φ(n), respectively.
d is kept as the private key exponent.
 
The public key consists of the modulus n and the public (or encryption) exponent e. The private key consists of the modulus n and the private (or decryption) exponent d, which must be kept secret. p, q, and φ(n) must also be kept secret because they can be used to calculate d.